+++ next.config.mjs
@@ -1,6 +1,4 @@
 /** @type {import('next').NextConfig} */
-const nextConfig = {
-    reactStrictMode: true,
-  };
+const nextConfig = {};
 
 export default nextConfig;
