+++ src/app/pokemons/page.tsx
@@ -1,0 +1,5 @@
+function page() {
+  return <div>page</div>;
+}
+
+export default page;
